:art: Split Accounts views and urls in respective domains
