+++ source/YugiAcademy/Accounts/user_profiles/views/__init__.py
@@ -1,0 +1,1 @@
+from .update_profile_view import update_profile
