+++ source/YugiAcademy/Accounts/users/views/update_password_view.py
@@ -1,0 +1,23 @@
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import redirect, render
+
+from users.forms import UpdatePasswordForm
+
+
+@login_required
+def update_password(request):
+    if request.method == 'POST':
+        form = UpdatePasswordForm(request.user, request.POST)
+        if form.is_valid():
+            form.save()
+            messages.info(request, "Password has been successfully updated")
+            # Django will automagically redirect to login
+        else:
+            messages.error(request, 'Password update has failed.')
+            return redirect('profile')
+
+    if request.method == 'GET':
+        return render(request, "update_password.html", {"form": UpdatePasswordForm(request.user)})
+
+    raise Exception(f'The {request.method} method was not expected')
